requirements.txt
numpy==1.14.3
opencv-contrib-python==3.4.1.15
opencv-python==3.4.1.15
Pillow==5.1.0
PyQt5==5.10.1
sip==4.19.8

